+++ admin/tabs/activity/tracking/detail/ActivityDetailModel.ts
@@ -32,7 +32,21 @@ export class ActivityDetailModel extends HoistModel {
             sortBy: 'dateCreated|desc',
             colChooserModel: true,
             enableExport: true,
-            filterModel: true,
+            filterModel: {
+                fieldSpecs: [
+                    'username',
+                    'category',
+                    'msg',
+                    'device',
+                    'browser',
+                    'userAgent',
+                    'elapsed',
+                    {
+                        field: 'dateCreated',
+                        fieldType: 'localDate'
+                    }
+                ]
+            },
             exportOptions: {
                 columns: 'ALL',
                 filename: exportFilename('activity-detail')