+++ cmp/filter/FilterChooserFieldSpec.ts
@@ -122,6 +122,11 @@ export class FilterChooserFieldSpec extends BaseFilterFieldSpec {
         if (!valueParser && (this.fieldType === 'int' || this.fieldType === 'number')) {
             return input => parseNumber(input);
         }
+
+        // Default date parser if fieldSpec's fieldType is localDate and record's field type is date.
+        if (!valueParser && this.filterDateAsLocalDate) {
+            return (v, op) => this.parseLocalDateAsDate(v, op);
+        }
         return valueParser;
     }